Better args handle, leak free

+++ libft/srcs/ft_atoi.c
@@ -12,7 +12,7 @@
 
 #include "../include/libft.h"
 
-unsigned long long		ft_atoi(const char *str)
+long		ft_atoi(const char *str)
 {
 	int		i;
 	int		sign;
